Atualização: correções

+++ matrizes.h
@@ -10,7 +10,7 @@ void preencher_mat_cript(int (*mat)[MAT]);
 int calcular_determinante(int (*mat)[MAT]);
 
 // calcula a matriz de descriptografia
-void preencher_mat_decript(int (*mat_cript)[MAT], float (*mat_decript)[MAT]);
+void preencher_mat_decript(int (*mat_cript)[MAT], int (*mat_decript)[MAT]);
 
 // imprime matrizes
 void imprimir_matriz_float(float (*mat)[MAT]);
